BO layer >> sabe in to arraylist and get All notes
