+++ src/main/java/lk/ijse/notetaker/service/NoteService.java
@@ -1,10 +1,10 @@
-package lk.ijse.notetaker.bo;
+package lk.ijse.notetaker.service;
 
 import lk.ijse.notetaker.dto.NoteDTO;
 
 import java.util.List;
 
-public sealed interface NoteBO permits NoteBOImpl{
+public sealed interface NoteService permits NoteServiceImpl {
 
     String saveData(NoteDTO noteDTO);
     boolean updateNote(String noteId,NoteDTO noteDTO);
